backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lore-backend"
version = "0.1.0"
description = "Lore — Organizational Memory Layer for AI-Native Companies"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",

    # Data validation
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",

    # PostgreSQL (Supabase)
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",

    # Neo4j graph DB — deferred to M6 (using PostgreSQL for graph queries at MVP)
    # "neo4j>=5.18.0",

    # Event streaming (Upstash Kafka)
    "aiokafka>=0.10.0",

    # Cache / KV (Upstash Redis)
    "redis[asyncio]>=5.0.0",

    # HTTP client
    "httpx>=0.27.0",

    # Auth (Clerk webhook verification)
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.9",

    # Structured logging
    "structlog>=24.1.0",

    # ID generation (ULID — time-ordered, URL-safe)
    "python-ulid>=1.0.3",

    # LLM providers
    "openai>=1.14.0",
    "anthropic>=0.20.0",
    "google-generativeai>=0.5.0",

    # Crypto (Slack/GitHub signature verification)
    "cryptography>=42.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.27.0",        # AsyncClient for test requests
    "ruff>=0.3.0",
    "mypy>=1.9.0",
    "faker>=24.0.0",        # Test data generation
]

[project.scripts]
lore-api = "app.main:app"

[tool.hatch.build.targets.wheel]
packages = ["app"]

# ── pytest ────────────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
env = [
    "ENVIRONMENT=testing",
    "DATABASE_URL=postgresql+asyncpg://postgres:postgres@localhost:5432/lore_test",
]

# ── ruff (linter + formatter) ─────────────────────────────────────────────────
[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

# ── mypy ──────────────────────────────────────────────────────────────────────
[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
